requirements.txt
fastapi==0.78.0
pydantic==1.9.0
uvicorn[standard]==0.17.6
configargparse==1.5.3
coloredlogs==15.0.1
paramiko==2.11.0
psycopg2-binary==2.9.3
